// app/schemas/message.ts
import z from "zod";

export const createMessageSchema = z.object({
  channelId: z.string(),
  content: z.string(),
  imageUrl: z.url().optional(),
  threadId: z.string().optional(),
});

export const updateMessageSchema = z.object({
  messageId: z.string(),
  content: z.string(),
});

export type CreateMessageSchemaType = z.infer<typeof createMessageSchema>
export type UpdateMessageSchemaType = z.infer<typeof updateMessageSchema>