// app/router/message.ts
import z from "zod";
import { heavyWriteSecurityMiddleware } from "../middlewares/arcjet/heavy-write";
import { standardSecurityMiddleware } from "../middlewares/arcjet/standard";
import { requiredAuthMiddleware } from "../middlewares/auth";
import { base } from "../middlewares/base";
import { requiredWorkspaceMiddleware } from "../middlewares/workspace";
import prisma from "@/lib/db";
import { createMessageSchema } from "../schemas/message";
import { getAvatar } from "@/lib/get-avatar";
import { Message } from "@/lib/generated/prisma/client";

export const createMessage = base
  .use(requiredAuthMiddleware)
  .use(requiredWorkspaceMiddleware)
  .use(standardSecurityMiddleware)
  .use(heavyWriteSecurityMiddleware)
  .route({
    method: "POST",
    path: "/messages",
    summary: "Create a message",
    tags: ["Messages"],
  })
  .input(createMessageSchema)
  .output(z.custom<Message>())
  .handler(async ({ input, context, errors }) => {
    // verify thye channel belongs to the user's organization
    const channel = await prisma.channel.findFirst({
      where: {
        id: input.channelId,
        workspaceId: context.workspace.orgCode,
      },
    });

    if (!channel) {
      throw errors.FORBIDDEN();
    }

    const created = await prisma.message.create({
      data: {
        content: input.content,
        imageUrl: input.imageUrl,
        channelId: input.channelId,
        authorId: context.user.id,
        authorEmail: context.user.email!,
        authorName: context.user.given_name!,
        authorAvatar: getAvatar(context.user.picture, context.user.email!)
      }
    });

    return {
      ...created,
    }
  });

  export const listMessages = base
    .use(requiredAuthMiddleware)
    .use(requiredWorkspaceMiddleware)
    .use(standardSecurityMiddleware)
    .use(heavyWriteSecurityMiddleware)
    .route({
      method: "GET",
      path: "/messages",
      summary: "List all messages",
      tags: ["Messages"],
    })
    .input(z.object({
      channelId: z.string(),
    }))
    .output(z.array(z.custom<Message>()))
    .handler(async ({ input, context, errors }) => {
      const channel = await prisma.channel.findFirst({
        where: {
          id: input.channelId,
          workspaceId: context.workspace.orgCode,
        },
      });

      if (!channel) {
        throw errors.FORBIDDEN();
      }

      const data = await prisma.message.findMany({
        where: {
          channelId: input.channelId,
        },
        orderBy: {
          createdAt: "desc",
        },
      });

      return data;
    });
